+++ scanpipe/pipes/d2d.py
@@ -24,6 +24,8 @@ from itertools import islice
 from pathlib import Path
 from timeit import default_timer as timer
 
+from django.db.models import Q
+
 from scanpipe import pipes
 from scanpipe.models import CodebaseRelation
 from scanpipe.models import CodebaseResource
@@ -41,17 +43,17 @@ TO = "to/"
 
 
 def get_inputs(project):
-    """Locate the ``from`` and ``to`` archives in project inputs directory."""
-    from_file = list(project.inputs("from*"))
-    to_file = list(project.inputs("to*"))
+    """Locate the ``from`` and ``to`` input files in project inputs/ directory."""
+    from_files = list(project.inputs("from*"))
+    to_files = list(project.inputs("to*"))
 
-    if len(from_file) != 1:
-        raise FileNotFoundError("from* archive not found.")
+    if len(from_files) < 1:
+        raise FileNotFoundError("from* input files not found.")
 
-    if len(to_file) != 1:
-        raise FileNotFoundError("to* archive not found.")
+    if len(to_files) < 1:
+        raise FileNotFoundError("to* input files not found.")
 
-    return from_file[0], to_file[0]
+    return from_files, to_files
 
 
 def get_resource_codebase_root(project, resource_path):
@@ -614,11 +616,22 @@ def _map_about_file_resource(project, about_file_resource, to_resources):
         # Cannot map anything without the about_resource value.
         return
 
+    ignored_resources = []
+    if extra_data := package_data.get("extra_data"):
+        ignored_resources = extra_data.get("ignored_resources")
+
     # Fetch all resources that are covered by the .ABOUT file.
     codebase_resources = to_resources.filter(path__contains=f"/{filename.lstrip('/')}")
     if not codebase_resources:
         # If there's nothing to map on the ``to/`` do not create the package.
         return
+
+    # Ignore resources for paths in `ignored_resources` attribute
+    if ignored_resources:
+        lookups = Q()
+        for resource_path in ignored_resources:
+            lookups |= Q(**{"path__contains": resource_path})
+        codebase_resources = codebase_resources.filter(~lookups)
 
     # Create the Package using .ABOUT data and assigned related codebase_resources
     pipes.update_or_create_package(project, package_data, codebase_resources)